requirements.txt
anyio==4.0.0
APScheduler==3.10.4
certifi==2023.7.22
colorama==0.4.6
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
py-mon==1.1.1
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
tzlocal==5.2
watchdog==3.0.0
